Admin UI: moved some files to more logical places
