+++ .changeset/gold-scissors-admire.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/app-admin-ui': patch
+---
+
+Moved a couple files around to more logical places.
